Tidy up and auto-format
