+++ spacy/lang/tn/stop_words.py
@@ -1,18 +1,20 @@
 # Stop words
-STOP_WORDS = set("""
-ke gareng ga selekanyo tlhwatlhwa yo mongwe se 
+STOP_WORDS = set(
+    """
+ke gareng ga selekanyo tlhwatlhwa yo mongwe se
 sengwe fa go le jalo gongwe ba na mo tikologong
-jaaka kwa morago nna gonne ka sa pele nako teng 
+jaaka kwa morago nna gonne ka sa pele nako teng
 tlase fela ntle magareng tsona feta bobedi kgabaganya
 moo gape kgatlhanong botlhe tsotlhe bokana e esi
-setseng mororo dinako golo kgolo nnye wena gago 
-o ntse ntle tla goreng gangwe mang yotlhe gore 
-eo yona tseraganyo eng ne sentle re rona thata 
-godimo fitlha pedi masomamabedi lesomepedi mmogo 
-tharo tseo boraro tseno yone jaanong bobona bona 
-lesome tsaya tsamaiso nngwe masomethataro thataro 
+setseng mororo dinako golo kgolo nnye wena gago
+o ntse ntle tla goreng gangwe mang yotlhe gore
+eo yona tseraganyo eng ne sentle re rona thata
+godimo fitlha pedi masomamabedi lesomepedi mmogo
+tharo tseo boraro tseno yone jaanong bobona bona
+lesome tsaya tsamaiso nngwe masomethataro thataro
 tsa mmatota tota sale thoko supa dira tshwanetse di mmalwa masisi
-bonala e tshwanang bogolo tsenya tsweetswee karolo 
-sepe tlhalosa dirwa robedi robongwe lesomenngwe gaisa 
-tlhano lesometlhano botlalo lekgolo           
-""".split())
+bonala e tshwanang bogolo tsenya tsweetswee karolo
+sepe tlhalosa dirwa robedi robongwe lesomenngwe gaisa
+tlhano lesometlhano botlalo lekgolo
+""".split()
+)
