+++ spacy/tests/matcher/test_phrase_matcher.py
@@ -327,8 +327,10 @@ def test_phrase_matcher_sent_start(en_vocab, attr):
 
 def test_span_in_phrasematcher(en_vocab):
     """Ensure that PhraseMatcher accepts Span and Doc as input"""
-    doc = Doc(en_vocab,
-              words=["I", "like", "Spans", "and", "Docs", "in", "my", "input", ",", "and", "nothing", "else", "."])
+    # fmt: off
+    words = ["I", "like", "Spans", "and", "Docs", "in", "my", "input", ",", "and", "nothing", "else", "."]
+    # fmt: on
+    doc = Doc(en_vocab, words=words)
     span = doc[:8]
     pattern = Doc(en_vocab, words=["Spans", "and", "Docs"])
     matcher = PhraseMatcher(en_vocab)
@@ -341,10 +343,14 @@ def test_span_in_phrasematcher(en_vocab):
 
 def test_span_v_doc_in_phrasematcher(en_vocab):
     """Ensure that PhraseMatcher only returns matches in input Span and not in entire Doc"""
-    doc = Doc(en_vocab,
-              words=["I", "like", "Spans", "and", "Docs", "in", "my", "input", ",", 
-                    "Spans", "and", "Docs", "in", "my", "matchers", ","
-                    "and", "Spans", "and", "Docs", "everywhere" "."])
+    # fmt: off
+    words = [
+        "I", "like", "Spans", "and", "Docs", "in", "my", "input", ",", "Spans",
+        "and", "Docs", "in", "my", "matchers", "," "and", "Spans", "and", "Docs",
+        "everywhere", "."
+    ]
+    # fmt: on
+    doc = Doc(en_vocab, words=words)
     span = doc[9:15]  # second clause
     pattern = Doc(en_vocab, words=["Spans", "and", "Docs"])
     matcher = PhraseMatcher(en_vocab)
