+++ spacy/tests/vocab_vectors/test_lexeme.py
@@ -57,6 +57,7 @@ def test_vocab_lexeme_add_flag_provided_id(en_vocab):
     assert en_vocab["dogs"].check_flag(is_len4) is True
     en_vocab.add_flag(lambda string: string.isdigit(), flag_id=IS_DIGIT)
 
+
 def test_vocab_lexeme_oov_rank(en_vocab):
     """Test that default rank is OOV_RANK."""
     lex = en_vocab["word"]
